refactor: move bisect.ts to common utils
